Merged revision 9110 from http://codespeak.net/svn/kupu/branch/kupu-1.2
Make Plone xml files put the portal charset in the XML header.
There doesn't appear to be a clean way to do this.
At the same time, set the response content/type to text/xml so that retrieving pages such as kupumyitems.xml will display the xml in the browser.