Fix the drop method emulation in sortables
we need to also change the order of elements programmatically
which messes up the internal state
so, we have a sortreset event that must be called up in this case
to recover state.
To get rid of this workaround, we'd need extensive customization
of prototype, and only possible with dupplicating code at the
moment.