Moved the auto selector converting (pass a string and it will be converted to a css selector) to the commands so that all commands will automatically have it.
Removed unneeded import from tests.py (was used for teardown)

Changed the core plugin config so that only the required extra
Javascript is loaded from the 3rd party directory. This should speed
up page loads.

Loading a non existing plugin now raises an exception. This means that
it is easier to tell if you made a type/forgot to install the plugin.