Add our own implementation of a language negotiator that actually uses the language cookie; the Zope 3 language negotiator wouldn't care about this.
Also, add an evil hack where we pretend that 'en' is always available;
this has the potential of making non human readable ids appear.

The problem we're trying to solve with this is that when you have 'zh'
as your second language, and 'en' as your first, we would normally
return the Chinese translation, which is awkward.  Another workaround
would be to provide an English translation PO file.  This should be
solved in a better way, really.