The HTML template now generates links to the KSS Javascript and KSS resource for a test.
Implemented the loading of the KSS Javascript. All Javascript include
the extras is concatinatinated into one big file.

Loading of KSS resources works.

Server handlers now work. There is an example of this in the core
plugin config.

Together all this means that we have all the (primitive) ingredients
for testing KSS without depending on a full webframework.