the extended browsertest was failing badly, since the [http://svn.zope.org/Zope/tags/2.10.5/lib/python/Shared/DC/ZRDB/TM.py?rev=81177&view=markup example mixin class providing a transaction manager hook] from [http://svn.zope.org/Zope/tags/2.10.5/lib/python/Shared/DC/ZRDB/?rev=81177 Shared.DC.ZRDB] on which the [https://svn.enfoldsystems.com/browse/public/enfold.solr/branches/snowsprint08-buildout/enfold.solr/enfold/solr/tm.py?rev=1792&view=markup transaction manager code] from [https://svn.enfoldsystems.com/browse/public/enfold.solr/branches/snowsprint08-buildout/enfold.solr/?rev=1792 enfold.solr][*] relied upon, might work for hooking up data managers that don't interact with the zodb, but is nevertheless plain broken...
all the work is done in `tpc_finish`, when — at least according to the [http://svn.zope.org/ZODB/tags/3.7.1/src/transaction/interfaces.py?rev=77348&view=markup IDataManager interface] — this should only be used to "indicate" that things have been made persistent and "should never fail".  and that's for a reason too, which is that when `tpc_finish` gets called any changes to zodb objects have already been committed.  that happens during the call to `commit` of the last hook registered with the TM, which is the connection to the zodb itself.  so trying to index stuff on the portal catalog at that point fails in very strange and interesting ways...

instead, the "real work" needs to be done in `commit()`, who's docstring also tells us so: "commit modifications to registered objects".  this way the changes automatically get committed as one would expect and everything is fine...  considering the time it took to debug this, i think the example / mixin class in `Shared.DC.ZRDB` should be changed asap! :(

[*] which funnily was about the only bit remaining from the original enfold packages after [browser:collective.indexing/trunk/README.txt#60682 this rewrite] of [http://lists.plone.org/pipermail/enterprise/2008-January/000001.html enfold's indexing approach]... %)