* check to see if the buildout is running under windows  * if running under windows make the following assumptions    * zeo server will be zeoservice.exe    * instance script will be <instance part name>.exe
This should allow zopectl scripts to be run under windows.  One issue I ran into was that the path to the script would contain strange characters and error out.  To get around the issue I defined script = bin/<partname>.py as I'm not sure if it was specific to my setup.