Found bug in getSubNavigationFoldersXML().
The function parameter must be a unicode string, but to get around apparent bug in formlib Choice() field not handling unicode properly (investigating that next!) our current workaround is to replace unicode with "??".

By replacing unicode with "??" this function can no longer find the top level item in the catalog, so returns no secondary items.