Don't check for the 'tramline' header in the request.
Even when tramline was intercepting uploads, files weren't being moved to the
repository directory because ATTramline was checking for a 'tramline' header in
the request before sending out a 'tramline_ok' header in the response.

There is apparently a bug in tramline which prevents the 'tramline' header
making it to the backend server.

Fixes http://plone.org/products/attramline/issues/3