Unit tests fixed.
Most of the unit tests were in the form:

  create -t foo.bar ...

And then would assume that the generated package would be called
"foo.bar", but the created folders would be "plone/example", since
those were the hardcoded recipe defaults.

Now that we split the project package name -> packages, this would
create "foo.bar/foo/bar".