fix parsing of "date" fields with Solr 1.4:  `Archetypes`'s "floor date" of `DateTime(1000,1)`, which is used as a fallback for indices like `effective`, is converted to different time representations depending on the architecture being 32- or 64-bit.
that's because for 32-bit the number of seconds since, or rather before, the epoch is bigger than `sys.maxint`, causing `DateTime.safegmtime` to raise a `TimeError`.  however, `safegmtime` (here used by `toZone`) catches that error, then calculates the time zone offset otherwise, finally ending up with a (correct) date from december 999 a.d.  the ISO8601 representation is then indexed in Solr, but version 1.4 omits any leading zero when returning the value as part of a search result, which breaks `DateTime`'s parser with:
{{{
 ...
 Module DateTime.DateTime, line 1013, in _parse
SyntaxError: 999-12-31T23:00:00Z
}}}

on a 64-bit machine the story is different:  here the seconds don't exceed `sys.maxint` and are therefore passed to `time.gmtime`.  for dates before the start of the [http://en.wikipedia.org/wiki/Gregorian_calendar gregorian calendar], `DateTime` gets the count wrong, though, resulting in an offset of five days:
{{{
>>> from DateTime import DateTime
>>> DateTime(1000, 1, 1).HTML4()
'1000-01-05T23:00:00Z'
}}}
that's clearly a bug, but one that doesn't matter too much regarding this particular use case:  the resulting date doesn't break Solr, and "floor dates" only need to be smaller/before any other date and are never compared to anything anyway...

all in all, the problem is specific to 32-bit machines in combination with Solr 1.4 and it can probably be blamed on the pope! :)