Respect Site Setup/Security/"Use email address as login name" by merging:
    http://svn.plone.org/svn/collective/Products.remember/branches/maurits-emaillogin-pas

and adjusting the operation so it respects the "Use email address as login
name" site property.

New accounts are not allowed to have the same email address as already
existing ones when the property is True, but the policy is not enforced
when it is false.  (Already existing accounts are not subject to the
constraint in either case.)

Currently the setup requires manual addition of the Remember Email User
Authentication plugin in acl_users, and enabling of the Extraction
authenticator once the plugin has been added.  I will make that automatic
with Products.remember (re)installation, next.