Do not set a default value during indexing.
When setting the contentgovernance default value, a changeOwnership is triggered, which then
causes a reindex. This is disastrous when this happens when indexing is already taking place,
such as when invoking the responsibleperson indexer. By setting _initializing_ to True we
prevent the default from being set.

Included is a view that'll clean up the resulting ZCatalog b0rkeness.