Let only the LDAP multi plugin implement the IGroupCapability interface
This is for adding a user to a group or removing a user from a group.
Previously the AD multi plugin claimed to implement this too, but it
lacked the required methods, so this could lead to tracebacks.  We
could instead add those methods via the mixin class, but but this gave
other tracebacks (in removePrincipalFromGroup) when I tried it in one
AD setup.